+++ src/App.js
@@ -3,7 +3,8 @@ import { withRouter } from 'react-router-dom';
 import styled, { ThemeProvider } from 'styled-components';
 import GlobalStyles from 'styles/global';
 import colorConverter from 'colorConverter';
-import { randomRgbValues, trueMod, hslTo4x } from 'helpers';
+import { trueMod } from 'helpers';
+import { randomRgbValues, hslTo4x } from 'colorUtils';
 import config from 'config';
 import { dark, light } from 'styles/themes';
 import breakpoints from 'styles/breakpoints';
