+++ src/components/HarmonyDisplay.js
@@ -3,7 +3,8 @@ import styled, { css } from 'styled-components';
 import colorConverter from 'colorConverter';
 import breakpoints from 'styles/breakpoints';
 import { useSpring, animated } from 'react-spring';
-import { trueMod, recordGAEvent, isBright } from 'helpers';
+import { trueMod, recordGAEvent } from 'helpers';
+import { isBright } from 'colorUtils';
 import config from 'config';
 import IconButton from 'components/common/IconButton';
 import { ReactComponent as MaxIcon } from 'icons/maximize.svg';
