+++ src/components/HslFields.js
@@ -2,7 +2,8 @@ import React, { useState, useEffect } from 'react';
 import ContiguousInputs from 'components/common/ContiguousInputs';
 import DegreeInput from 'components/common/DegreeInput';
 import HectoInput from 'components/common/HectoInput';
-import { recordGAEvent, hslTo4x } from 'helpers';
+import { recordGAEvent } from 'helpers';
+import { hslTo4x } from 'colorUtils';
 
 const HslFields = ({ setColor, colorValues }) => {
   const [hslValues, setHslValues] = useState({ h: 0, s: 0, l: 0 });
